+++ examples/housing_different_levels.py
@@ -42,9 +42,13 @@ f.manual_forecast(how='weighted',models=level_models,determine_best_by='Validati
 # plot results
 matplotlib.use('QT5Agg')
 f.plot(models='all',order_by='LevelTestSetMAPE',print_attr=['LevelTestSetRMSE','LevelTestSetR2','LevelTestSetMAPE','HyperParams','Xvars','models']) # will automatically plot levels for everything
+plt.show()
 f.plot_test_set(models='all',order_by='LevelTestSetMAPE',include_train=60) # will automatically plot levels for everything
+plt.show()
 f.plot_fitted(models=models,order_by='LevelTestSetMAPE') # cannot plot fitted values of all models when levels were different during forecasting
+plt.show()
 f.plot_fitted(models=level_models,order_by='LevelTestSetMAPE')
+plt.show()
 
 # export key results
 f.export(['model_summaries','lvl_fcsts'],to_excel=True,determine_best_by='LevelTestSetMAPE',excel_name='housing_different_levels_results.xlsx') # export interesting model metrics and forecasts (both level and non-level)